backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
httpx>=0.28.1
google-generativeai==0.3.2
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.0
poppler-utils